Fixed a compiler warning.
git-svn-id: https://svn.code.sf.net/p/quark/code@9412 5419a3ea-6bc3-475c-b79b-167d7c3fbd5f

+++ source/trunk/python/PyForms.pas
@@ -1045,7 +1045,7 @@ begin
     end
    else
     Info.wID:=Msg.wParamLo;  { single item }
-   if (Info.wID>0) and (Info.wID <= Callbacks.Count) then
+   if (Info.wID > 0) and (Info.wID <= UINT(Callbacks.Count)) then
     begin
      obj:=PyObject(Callbacks[Info.wID-1]);
      if obj<>Nil then
@@ -1081,7 +1081,7 @@ begin
  else
   I:=CurrentPopupHandle;
  if not FindMenuItem(I, Msg.wParam, Info) then Exit;
- if (Info.wID<=0) or (Info.wID>Callbacks.Count) then
+ if (Info.wID <= 0) or (Info.wID > UINT(Callbacks.Count)) then
   begin
    if CurrentPopupHandle<>0 then Exit;
    if Info.wID = wID_HelpMenu then
